==> experiments/02-region-locus-simulate/06-deming.ipynb <==
import numpy as np
import admix_genet_cor

v1 = v2 = np.random.normal(size=1000)
cor = 1.0
v = np.random.multivariate_normal(mean=[0, 0], cov=[[1.0, cor], [cor, 1.0]], size=1000)
v1, v2 = v[:, 0], v[:, 1]

np.random.seed(1234)
slope_list = []
for _ in range(100):
    x = v1 + np.random.normal(size=1000, scale=2.0)
    y = v2 + np.random.normal(size=1000, scale=0.9)
    slope = admix_genet_cor.locus.deming_regression(
        x=x,
        y=y,
        sx=2.0,
        sy=1.0,
        no_intercept=True,
    )
#     slope = admix_genet_cor.locus.deming_regression(
#         x=y,
#         y=x,
#         sx=1.0,
#         sy=2.0,
#         no_intercept=True,
#     )
        
    slope_list.append(slope)
print(np.mean(slope_list), np.std(slope_list) / np.sqrt(len(slope_list)))

import pandas as pd
pd.DataFrame({"x": x, "y": y}).to_csv("tmp-data.csv", index=False)

slope1 = admix_genet_cor.locus.deming_regression(
    x=x,
    y=y,
    sx=10.0,
    sy=5.0,
    no_intercept=False,
)
print(slope1)
slope2 = admix_genet_cor.locus.deming_regression(
    x=y,
    y=x,
    sx=5.0,
    sy=10.0,
    no_intercept=False,
)
print(slope2)

slope2

slope_list = []
for _ in range(100):
    slope = admix_genet_cor.locus.deming_regression(
        x=v1 + np.random.normal(size=1000, scale=5.0),
        y=v2 + np.random.normal(size=1000, scale=10.0),
        sx=1.0,
        sy=2.0,
        no_intercept=True,
    )
    slope_list.append(slope)
print(np.mean(slope_list), np.std(slope_list) / np.sqrt(len(slope_list)))
